refactor: split runner service command module
