+++ crates/runner/src/cmd/service/gate.rs
@@ -1,0 +1,543 @@
+use std::path::{Path, PathBuf};
+
+use crate::error::{ActiveJobsError, RunnerError, RunnerResult};
+use crate::ids::RunId;
+use crate::paths::HomePaths;
+use tracing::{info, warn};
+
+use super::diagnostic::status_field_preview;
+use super::systemctl::is_unit_active;
+use super::target::RunnerServiceUnit;
+
+/// Resolve the runner's base_dir from its service name suffix using the
+/// project-wide convention: `/var/lib/vm0-runner/runners/<suffix>/`.
+///
+/// This matches `ansible/playbooks/build-runner.yml` and the `--runner-dirname`
+/// default in `runner config`. Non-standard `base_dir` overrides (dev-only)
+/// will fail to locate status.json and fall through to forceful stop.
+pub(super) fn runner_base_dir(unit: &RunnerServiceUnit) -> Option<PathBuf> {
+    let home = HomePaths::new().ok()?;
+    Some(home.runners_dir().join(unit.suffix()))
+}
+
+/// Parsed snapshot of the runner's status.json.
+pub(super) struct RunnerStatusSnapshot {
+    /// Mode string sourced verbatim from status.json. Valid values are the
+    /// lowercase serialization of [`crate::status::RunnerMode`]: `"running"`,
+    /// `"draining"`, `"stopped"`. Unknown values (e.g. from a newer runner
+    /// writing a future variant) are preserved and routed to the normal
+    /// refuse branch by [`decide_gate`].
+    pub(super) mode: String,
+    /// UUIDs of runs currently in flight.
+    run_ids: Vec<RunId>,
+    /// How long the runner process itself has been up, derived from the
+    /// `started_at` timestamp. status.json does not record per-run start
+    /// times, so the error message surfaces this runner-level uptime
+    /// rather than a misleading per-job duration.
+    uptime: std::time::Duration,
+}
+
+#[derive(Debug)]
+pub(super) enum RunnerStatusReadError {
+    Read {
+        path: PathBuf,
+        error: std::io::Error,
+    },
+    ParseJson {
+        path: PathBuf,
+        error: serde_json::Error,
+    },
+    ParseStartedAt {
+        started_at: String,
+        error: chrono::ParseError,
+    },
+}
+
+impl std::fmt::Display for RunnerStatusReadError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::Read { path, error } => write!(f, "read {}: {error}", path.display()),
+            Self::ParseJson { path, error } => write!(f, "parse {}: {error}", path.display()),
+            Self::ParseStartedAt { started_at, error } => {
+                write!(
+                    f,
+                    "parse started_at {:?}: {error}",
+                    status_field_preview(started_at)
+                )
+            }
+        }
+    }
+}
+
+/// Decision from [`decide_gate`] — pure function that maps a status
+/// snapshot to the gate outcome without performing any I/O.
+#[derive(Debug, PartialEq, Eq)]
+enum GateDecision {
+    /// Let the stop/uninstall proceed.
+    Bypass,
+    /// Refuse the operation; `draining` selects the UX variant.
+    Refuse { draining: bool },
+}
+
+/// Pure decision logic shared by the gate — testable without systemctl.
+///
+/// Short-circuit order:
+/// 1. `mode == "stopped"` or `"stopping"` — teardown has already started.
+///    The runner is actively cancelling any in-flight jobs itself; a user
+///    `stop`/`uninstall` just accelerates the process and is safe.
+/// 2. `run_ids.is_empty()` — nothing to protect, regardless of mode.
+/// 3. Otherwise refuse; `draining=true` when `mode == "draining"` so the
+///    error message suggests waiting rather than re-running `drain`.
+///
+/// Mode strings mirror [`crate::status::RunnerMode`] (serde lowercase).
+fn decide_gate(status: &RunnerStatusSnapshot) -> GateDecision {
+    if matches!(status.mode.as_str(), "stopped" | "stopping") {
+        return GateDecision::Bypass;
+    }
+    if status.run_ids.is_empty() {
+        return GateDecision::Bypass;
+    }
+    GateDecision::Refuse {
+        draining: status.mode == "draining",
+    }
+}
+
+/// Read status.json at `base_dir`.
+///
+/// Returns an error on any I/O or parse failure — the caller should log the
+/// reason and fall through to forceful stop (status unknown → cannot protect
+/// jobs).
+async fn read_private_status_json(path: &Path) -> std::io::Result<String> {
+    match crate::private_fs::read_private_file_to_string_with_max(
+        path,
+        crate::private_fs::PRIVATE_STATUS_FILE_READ_MAX_BYTES,
+    )
+    .await
+    {
+        Ok(Some(content)) => Ok(content),
+        Ok(None) => Err(std::io::Error::new(
+            std::io::ErrorKind::NotFound,
+            format!("{} not found", path.display()),
+        )),
+        Err(error) => Err(std::io::Error::other(error.to_string())),
+    }
+}
+
+pub(super) async fn read_runner_status(
+    base_dir: &Path,
+) -> Result<RunnerStatusSnapshot, RunnerStatusReadError> {
+    #[derive(serde::Deserialize)]
+    struct StatusFile {
+        mode: String,
+        active_runs: Vec<ActiveRunEntry>,
+        started_at: String,
+    }
+    #[derive(serde::Deserialize)]
+    struct ActiveRunEntry {
+        run_id: RunId,
+        // `sandbox_id` is present in the file but unused by the gate.
+    }
+    let path = base_dir.join("status.json");
+    let content =
+        read_private_status_json(&path)
+            .await
+            .map_err(|error| RunnerStatusReadError::Read {
+                path: path.clone(),
+                error,
+            })?;
+    let file: StatusFile = serde_json::from_str(&content)
+        .map_err(|error| RunnerStatusReadError::ParseJson { path, error })?;
+    let started = chrono::DateTime::parse_from_rfc3339(&file.started_at).map_err(|error| {
+        RunnerStatusReadError::ParseStartedAt {
+            started_at: file.started_at.clone(),
+            error,
+        }
+    })?;
+    let now = chrono::Utc::now();
+    let uptime = (now - started.with_timezone(&chrono::Utc))
+        .to_std()
+        .unwrap_or_default();
+    Ok(RunnerStatusSnapshot {
+        mode: file.mode,
+        run_ids: file.active_runs.into_iter().map(|r| r.run_id).collect(),
+        uptime,
+    })
+}
+
+/// Gate for `service stop` / `service uninstall`: block the operation when
+/// the runner has active jobs unless `force` is set.
+///
+/// Returns `Ok(())` to proceed (either bypassed or confirmed safe). Returns
+/// `Err(RunnerError::ActiveJobs(_))` to refuse with a user-facing message.
+///
+/// ## Transient / race handling
+///
+/// Each of these conditions returns `Ok(())` to let the operator through,
+/// erring on the side of "stop is usable" over "gate is strict":
+///
+/// 1. **Dead / crashed runner** — if the systemd unit is inactive, the
+///    on-disk `active_runs` may be stale (runner was SIGKILLed before it
+///    could update status.json). Nothing alive to protect; skip the gate.
+/// 2. **Cleanly stopped runner** — `mode == "stopped"` indicates the
+///    runner's own drain finished. Covers the short window between
+///    status.json being rewritten with `"stopped"` (`start.rs` end of
+///    `run_with_config`) and systemd noticing the process has exited
+///    (marking the unit inactive). Without this, the gate could spuriously
+///    refuse a stop issued during that window.
+/// 3. **Base-dir unresolvable** — non-standard deployments that override
+///    `base_dir` away from the `/var/lib/vm0-runner/runners/<suffix>`
+///    convention fall here. Warn-log and fall through.
+/// 4. **Status file unreadable / malformed** — missing file, permission
+///    denied, JSON parse error, bad `started_at`: warn-log and fall
+///    through. Matches the acceptance criteria.
+///
+/// When the runner's `mode == "draining"`, we still refuse but flip the
+/// `draining` flag so the error renders a wait-or-force message (the
+/// operator already initiated drain, so re-suggesting drain would be
+/// noise).
+///
+/// ## TOCTOU (documented, not mitigated)
+///
+/// Between this gate reading status.json and `systemctl stop` killing the
+/// process, the runner may claim a new job via its API poll (seconds
+/// cadence). That job will be killed. This is *intentional*: `stop` is
+/// defined as forceful. Callers who need zero-race graceful shutdown
+/// should use `service drain`. Mitigating this race would require sending
+/// SIGUSR1 first and waiting — which is exactly what `drain` does.
+pub(super) async fn check_active_jobs_gate(
+    unit: &RunnerServiceUnit,
+    force: bool,
+    command_name: &'static str,
+) -> RunnerResult<()> {
+    if force {
+        // Leave an audit trail — --force is valid but destructive, so
+        // journalctl should show it was used if jobs later appear lost.
+        info!(
+            unit = %unit.unit_name(),
+            command = command_name,
+            "--force passed, bypassing active-jobs gate"
+        );
+        return Ok(());
+    }
+
+    // (1) Dead-runner short-circuit.
+    let active = is_unit_active(unit).await.unwrap_or_else(|e| {
+        warn!(unit = %unit.unit_name(), error = %e, "cannot check unit state — skipping active-jobs gate");
+        false
+    });
+    if !active {
+        return Ok(());
+    }
+
+    let Some(base_dir) = runner_base_dir(unit) else {
+        warn!(
+            unit = %unit.unit_name(),
+            "cannot determine vm0-runner home — skipping active-jobs gate"
+        );
+        return Ok(());
+    };
+    let status = match read_runner_status(&base_dir).await {
+        Ok(status) => status,
+        Err(e) => {
+            warn!(
+                unit = %unit.unit_name(),
+                base_dir = %base_dir.display(),
+                error = %e,
+                "cannot read status.json — skipping active-jobs gate"
+            );
+            return Ok(());
+        }
+    };
+
+    match decide_gate(&status) {
+        GateDecision::Bypass => Ok(()),
+        GateDecision::Refuse { draining } => Err(RunnerError::ActiveJobs(ActiveJobsError {
+            unit: unit.unit_name().to_string(),
+            suffix: unit.suffix().to_string(),
+            run_ids: status.run_ids,
+            runner_uptime: status.uptime,
+            command_name,
+            draining,
+        })),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    // -----------------------------------------------------------------
+    // status.json reader
+    // -----------------------------------------------------------------
+
+    #[tokio::test]
+    async fn read_runner_status_missing_file() {
+        let dir = tempfile::tempdir().unwrap();
+        assert!(read_runner_status(dir.path()).await.is_err());
+    }
+
+    #[tokio::test]
+    async fn read_runner_status_empty_json() {
+        let dir = tempfile::tempdir().unwrap();
+        tokio::fs::write(dir.path().join("status.json"), "{}")
+            .await
+            .unwrap();
+        // Missing required fields -> parse error.
+        assert!(read_runner_status(dir.path()).await.is_err());
+    }
+
+    #[tokio::test]
+    async fn read_runner_status_malformed_started_at() {
+        let dir = tempfile::tempdir().unwrap();
+        let s = r#"{"mode":"running","active_runs":[],"started_at":"not-a-date"}"#;
+        tokio::fs::write(dir.path().join("status.json"), s)
+            .await
+            .unwrap();
+        assert!(read_runner_status(dir.path()).await.is_err());
+    }
+
+    #[cfg(unix)]
+    #[tokio::test]
+    async fn read_runner_status_rejects_fifo_without_blocking() {
+        use std::ffi::CString;
+        use std::os::unix::ffi::OsStrExt;
+
+        let dir = tempfile::tempdir().unwrap();
+        let path = dir.path().join("status.json");
+        let c_path = CString::new(path.as_os_str().as_bytes()).unwrap();
+        let result = unsafe { libc::mkfifo(c_path.as_ptr(), 0o600) };
+        assert_eq!(
+            result,
+            0,
+            "mkfifo failed: {}",
+            std::io::Error::last_os_error()
+        );
+
+        let result = tokio::time::timeout(
+            std::time::Duration::from_secs(1),
+            read_runner_status(dir.path()),
+        )
+        .await;
+
+        assert!(result.is_ok(), "FIFO read should not block");
+        assert!(result.unwrap().is_err(), "FIFO status should be rejected");
+    }
+
+    #[tokio::test]
+    async fn read_runner_status_malformed_long_started_at_error_is_bounded() {
+        let dir = tempfile::tempdir().unwrap();
+        let started_at = "x".repeat(512);
+        let s = format!(r#"{{"mode":"running","active_runs":[],"started_at":"{started_at}"}}"#);
+        tokio::fs::write(dir.path().join("status.json"), s)
+            .await
+            .unwrap();
+
+        let err = match read_runner_status(dir.path()).await {
+            Ok(_) => panic!("expected malformed started_at to fail"),
+            Err(err) => err,
+        };
+        let message = err.to_string();
+
+        assert!(message.contains(&"x".repeat(128)));
+        assert!(message.contains("...[truncated]"));
+        assert!(!message.contains(&"x".repeat(129)));
+    }
+
+    #[tokio::test]
+    async fn read_runner_status_running_no_jobs() {
+        let dir = tempfile::tempdir().unwrap();
+        let s = r#"{"mode":"running","active_runs":[],"started_at":"2026-04-13T00:00:00.000Z"}"#;
+        tokio::fs::write(dir.path().join("status.json"), s)
+            .await
+            .unwrap();
+        let status = read_runner_status(dir.path()).await.unwrap();
+        assert_eq!(status.mode, "running");
+        assert!(status.run_ids.is_empty());
+    }
+
+    #[tokio::test]
+    async fn read_runner_status_with_active_jobs() {
+        let dir = tempfile::tempdir().unwrap();
+        let s = r#"{
+            "mode":"running",
+            "active_runs":[
+                {"run_id":"0191c4e0-0000-7000-8000-000000000001","sandbox_id":"aaaaaaaa-0000-7000-8000-000000000001"},
+                {"run_id":"0191c4e0-0000-7000-8000-000000000002","sandbox_id":"aaaaaaaa-0000-7000-8000-000000000002"}
+            ],
+            "started_at":"2026-04-13T00:00:00.000Z"
+        }"#;
+        tokio::fs::write(dir.path().join("status.json"), s)
+            .await
+            .unwrap();
+        let status = read_runner_status(dir.path()).await.unwrap();
+        assert_eq!(status.mode, "running");
+        assert_eq!(status.run_ids.len(), 2);
+    }
+
+    #[tokio::test]
+    async fn read_runner_status_draining_mode() {
+        let dir = tempfile::tempdir().unwrap();
+        let s = r#"{
+            "mode":"draining",
+            "active_runs":[
+                {"run_id":"0191c4e0-0000-7000-8000-000000000001","sandbox_id":"aaaaaaaa-0000-7000-8000-000000000001"}
+            ],
+            "started_at":"2026-04-13T00:00:00.000Z"
+        }"#;
+        tokio::fs::write(dir.path().join("status.json"), s)
+            .await
+            .unwrap();
+        let status = read_runner_status(dir.path()).await.unwrap();
+        assert_eq!(status.mode, "draining");
+        assert_eq!(status.run_ids.len(), 1);
+    }
+
+    #[tokio::test]
+    async fn read_runner_status_full_runner_payload() {
+        // Guard against schema drift: status.json written by StatusTracker
+        // (crates/runner/src/status.rs) contains more fields than the ones
+        // we care about. The decoder must tolerate the full payload.
+        let dir = tempfile::tempdir().unwrap();
+        let s = r#"{
+            "mode": "running",
+            "max_concurrent": 4,
+            "active_runs": [
+                {
+                    "run_id":"0191c4e0-0000-7000-8000-000000000001",
+                    "sandbox_id":"aaaaaaaa-0000-7000-8000-000000000001",
+                    "phase":"preparing",
+                    "phase_started_at":"2026-04-13T00:00:01.000Z"
+                },
+                {
+                    "run_id":"0191c4e0-0000-7000-8000-000000000002",
+                    "sandbox_id":"aaaaaaaa-0000-7000-8000-000000000002",
+                    "phase":"running",
+                    "phase_started_at":"2026-04-13T00:00:02.000Z"
+                }
+            ],
+            "idle_vms": [
+                {"session_id":"sess-1","sandbox_id":"bbbbbbbb-0000-7000-8000-000000000001"}
+            ],
+            "proxy_port": 8080,
+            "dns_port": 5300,
+            "started_at": "2026-04-13T00:00:00.000Z",
+            "updated_at": "2026-04-13T00:05:00.000Z"
+        }"#;
+        tokio::fs::write(dir.path().join("status.json"), s)
+            .await
+            .unwrap();
+        let status = read_runner_status(dir.path()).await.unwrap();
+        assert_eq!(status.mode, "running");
+        assert_eq!(status.run_ids.len(), 2);
+    }
+
+    #[tokio::test]
+    async fn read_runner_status_future_started_at_yields_zero_uptime() {
+        // Clock skew guard: if started_at is in the future (NTP correction,
+        // misconfigured clock), `to_std()` fails and we fall back to
+        // Duration::ZERO rather than propagating an error.
+        let dir = tempfile::tempdir().unwrap();
+        let future = (chrono::Utc::now() + chrono::Duration::hours(1))
+            .format("%Y-%m-%dT%H:%M:%S%.3fZ")
+            .to_string();
+        let s = format!(r#"{{"mode":"running","active_runs":[],"started_at":"{future}"}}"#);
+        tokio::fs::write(dir.path().join("status.json"), s)
+            .await
+            .unwrap();
+        let status = read_runner_status(dir.path()).await.unwrap();
+        assert_eq!(status.uptime, std::time::Duration::ZERO);
+    }
+
+    #[tokio::test]
+    async fn read_runner_status_started_at_without_millis() {
+        // StatusTracker writes millisecond precision today, but RFC 3339
+        // allows second-precision too. Make sure we accept both so a
+        // future format change won't silently break the gate.
+        let dir = tempfile::tempdir().unwrap();
+        let s = r#"{"mode":"running","active_runs":[],"started_at":"2026-04-13T00:00:00Z"}"#;
+        tokio::fs::write(dir.path().join("status.json"), s)
+            .await
+            .unwrap();
+        let status = read_runner_status(dir.path()).await.unwrap();
+        assert_eq!(status.mode, "running");
+    }
+
+    #[tokio::test]
+    async fn read_runner_status_stopped_preserves_ids() {
+        // Reader is unopinionated: it returns what's on disk, and the gate
+        // consults `mode` to decide whether to short-circuit.
+        let dir = tempfile::tempdir().unwrap();
+        let s = r#"{
+            "mode":"stopped",
+            "active_runs":[
+                {"run_id":"0191c4e0-0000-7000-8000-000000000001","sandbox_id":"aaaaaaaa-0000-7000-8000-000000000001"}
+            ],
+            "started_at":"2026-04-13T00:00:00.000Z"
+        }"#;
+        tokio::fs::write(dir.path().join("status.json"), s)
+            .await
+            .unwrap();
+        let status = read_runner_status(dir.path()).await.unwrap();
+        assert_eq!(status.mode, "stopped");
+        assert_eq!(status.run_ids.len(), 1);
+    }
+
+    // -----------------------------------------------------------------
+    // decide_gate — pure decision function
+    // -----------------------------------------------------------------
+
+    fn snapshot(mode: &str, run_count: usize) -> RunnerStatusSnapshot {
+        RunnerStatusSnapshot {
+            mode: mode.to_string(),
+            run_ids: (0..run_count).map(|_| RunId::nil()).collect(),
+            uptime: std::time::Duration::from_secs(600),
+        }
+    }
+
+    #[test]
+    fn decide_gate_running_with_jobs_refuses_normal() {
+        assert_eq!(
+            decide_gate(&snapshot("running", 3)),
+            GateDecision::Refuse { draining: false }
+        );
+    }
+
+    #[test]
+    fn decide_gate_running_without_jobs_bypasses() {
+        assert_eq!(decide_gate(&snapshot("running", 0)), GateDecision::Bypass);
+    }
+
+    #[test]
+    fn decide_gate_stopped_bypasses_even_with_stale_ids() {
+        // Covers the narrow window between status.json being rewritten
+        // with "stopped" and systemd marking the unit inactive.
+        assert_eq!(decide_gate(&snapshot("stopped", 2)), GateDecision::Bypass);
+    }
+
+    #[test]
+    fn decide_gate_stopping_bypasses() {
+        // Stopping = teardown in progress. The runner is already cancelling
+        // in-flight jobs itself; user stop/uninstall accelerates rather than
+        // endangers — bypass the active-jobs gate.
+        assert_eq!(decide_gate(&snapshot("stopping", 3)), GateDecision::Bypass);
+    }
+
+    #[test]
+    fn decide_gate_draining_with_jobs_refuses_draining_variant() {
+        assert_eq!(
+            decide_gate(&snapshot("draining", 1)),
+            GateDecision::Refuse { draining: true }
+        );
+    }
+
+    #[test]
+    fn decide_gate_unknown_mode_with_jobs_refuses_normal() {
+        // Forward-compat: a newer runner writing a mode string we don't
+        // recognize (e.g. "paused") gets the normal refuse branch —
+        // safer than bypassing, and does not impersonate "draining".
+        assert_eq!(
+            decide_gate(&snapshot("paused", 1)),
+            GateDecision::Refuse { draining: false }
+        );
+    }
+}
